# api/openapi.yaml
openapi: 3.0.3
info:
  title: Opero API
  version: 0.1.0
  description: >
    Opero people-operations API. This spec is the single source of truth —
    every operation is defined here before it is implemented. Operational
    routes like /health are infrastructure and intentionally not part of the
    contract.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /locations:
    get:
      operationId: listLocations
      tags: [roster]
      summary: List locations.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The tenant's locations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createLocation
      tags: [roster]
      summary: Create a location.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationRequest'
      responses:
        '201':
          description: Location created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /locations/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      operationId: getLocation
      tags: [roster]
      summary: Get a location by id.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The location.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateLocation
      tags: [roster]
      summary: Update a location.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationRequest'
      responses:
        '200':
          description: Updated location.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteLocation
      tags: [roster]
      summary: Delete a location.
      description: Deletes a location. Shifts at it have their location cleared (set to null).
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /shifts:
    get:
      operationId: listShifts
      tags: [roster]
      summary: List shifts.
      description: >
        Lists shifts, optionally filtered by employee, status, and a
        [from, to) window on the shift start time.
      security:
        - bearerAuth: []
      parameters:
        - name: employee_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, published]
        - name: from
          in: query
          required: false
          description: Inclusive lower bound on starts_at (ISO-8601).
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Exclusive upper bound on starts_at (ISO-8601).
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Matching shifts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shift'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createShift
      tags: [roster]
      summary: Create a shift (starts as a draft).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShiftRequest'
      responses:
        '201':
          description: Shift created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /shifts/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      operationId: getShift
      tags: [roster]
      summary: Get a shift by id.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The shift.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateShift
      tags: [roster]
      summary: Update a shift.
      description: Updates shift fields. Status transitions go through the publish endpoint, not here.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShiftRequest'
      responses:
        '200':
          description: Updated shift.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteShift
      tags: [roster]
      summary: Delete a shift.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /shifts/{id}/publish:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      operationId: publishShift
      tags: [roster]
      summary: Publish a shift.
      description: Transitions a shift to published, making it visible to the assigned staff. Idempotent.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The published shift.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /me/shifts:
    get:
      operationId: listMyShifts
      tags: [roster]
      summary: List the authenticated employee's own shifts.
      description: >
        Field-facing: returns shifts for the employee linked to the caller's
        login (resolved from the token), optionally filtered by status and a
        [from, to) window on start time. Returns an empty list if the caller has
        no linked employee record.
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, published]
        - name: from
          in: query
          required: false
          description: Inclusive lower bound on starts_at (ISO-8601).
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Exclusive upper bound on starts_at (ISO-8601).
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: The caller's shifts, ordered by start time.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shift'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication failed or is missing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: A stable, machine-readable error code.
          example: invalid_credentials
        message:
          type: string
          description: A human-readable explanation.
...
    Location:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
          nullable: true
        lat:
          type: number
          format: double
          nullable: true
        lng:
          type: number
          format: double
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateLocationRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        address:
          type: string
          nullable: true
        lat:
          type: number
          format: double
          nullable: true
        lng:
          type: number
          format: double
          nullable: true
    UpdateLocationRequest:
      type: object
      description: PATCH semantics — only provided fields are changed; fields cannot be cleared to null in v1.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        address:
          type: string
          nullable: true
        lat:
          type: number
          format: double
          nullable: true
        lng:
          type: number
          format: double
          nullable: true
    Shift:
      type: object
      required: [id, employee_id, starts_at, ends_at, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        location_id:
          type: string
          format: uuid
          nullable: true
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateShiftRequest:
      type: object
      required: [employee_id, starts_at, ends_at]
      properties:
        employee_id:
          type: string
          format: uuid
        location_id:
          type: string
          format: uuid
          nullable: true
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
    UpdateShiftRequest:
      type: object
      description: >
        PATCH semantics — only provided fields are changed; fields cannot be
        cleared to null in v1. Status transitions go through POST /shifts/{id}/publish.
      properties:
        employee_id:
          type: string
          format: uuid
        location_id:
          type: string
          format: uuid
          nullable: true
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
...